# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oogly Chat API specification
 * OpenAPI spec version: 1.4.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 32
     */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  isAdmin: boolean;
  isOwner: boolean;
  isMuted: boolean;
  /** @nullable */
  theme?: string | null;
  /** @nullable */
  notificationsEnabled?: boolean | null;
  /** @nullable */
  soundsEnabled?: boolean | null;
  /** @nullable */
  compactMode?: boolean | null;
  /** @nullable */
  lastSeen?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserUpdate {
  isAdmin?: boolean;
  isOwner?: boolean;
}

export interface UserListResponse {
  users: User[];
  total: number;
}

export interface SettingsUpdate {
  theme?: string;
  notificationsEnabled?: boolean;
  soundsEnabled?: boolean;
  compactMode?: boolean;
}

export interface ReactionSummary {
  emoji: string;
  count: number;
  userReacted: boolean;
}

export interface Message {
  id: number;
  content: string;
  authorId: number;
  authorUsername: string;
  deleted: boolean;
  /** @nullable */
  replyToId?: number | null;
  /** @nullable */
  replyToContent?: string | null;
  /** @nullable */
  replyToUsername?: string | null;
  reactions?: ReactionSummary[];
  createdAt: string;
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
  replyToId?: number;
}

export interface Reaction {
  id: number;
  messageId: number;
  userId: number;
  emoji: string;
  createdAt: string;
}

export interface ReactionInput {
  emoji: string;
}

export interface Ban {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  bannedById?: number | null;
  /** @nullable */
  bannedByUsername?: string | null;
  reason: string;
  isPermanent: boolean;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export interface BanInput {
  userId: number;
  /** @minLength 1 */
  reason: string;
  isPermanent?: boolean;
  /** @nullable */
  expiresAt?: string | null;
}

export interface BanStatus {
  isBanned: boolean;
  ban?: Ban;
}

export interface Mute {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  mutedById?: number | null;
  /** @nullable */
  mutedByUsername?: string | null;
  reason: string;
  isPermanent: boolean;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export interface MuteInput {
  userId: number;
  /** @minLength 1 */
  reason: string;
  isPermanent?: boolean;
  /** @nullable */
  expiresAt?: string | null;
}

export interface Appeal {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  message: string;
  isRead: boolean;
  isDismissed: boolean;
  createdAt: string;
}

export interface AppealInput {
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
}

export interface Announcement {
  id: number;
  content: string;
  authorId: number;
  /** @nullable */
  authorUsername?: string | null;
  createdAt: string;
}

export interface AnnouncementInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}

/**
 * @nullable
 */
export type ModerationLogMetadata = { [key: string]: unknown } | null;

export interface ModerationLog {
  id: number;
  /** @nullable */
  adminId?: number | null;
  adminUsername: string;
  /** @nullable */
  targetUserId?: number | null;
  /** @nullable */
  targetUsername?: string | null;
  action: string;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  metadata?: ModerationLogMetadata;
  createdAt: string;
}

export interface ModerationLogListResponse {
  logs: ModerationLog[];
  total: number;
}

export interface AdminStats {
  totalUsers: number;
  totalMessages: number;
  activeBans: number;
  activeMutes: number;
  pendingAppeals: number;
  messagesLast24h: number;
  newUsersLast24h: number;
}

export interface FilterCheckInput {
  text: string;
}

export interface FilterCheckResult {
  flagged: boolean;
  matches: string[];
  cleaned: string;
}

export type ListMessagesParams = {
/**
 * Cursor for pagination (message id)
 */
before?: number;
/**
 * Max number of messages to return
 */
limit?: number;
};

export type ListUsersParams = {
search?: string;
page?: number;
};

export type ListBansParams = {
active?: boolean;
};

export type ListMutesParams = {
active?: boolean;
};

export type ListAppealsParams = {
unread?: boolean;
};

export type ListModerationLogsParams = {
page?: number;
action?: string;
adminId?: number;
targetUserId?: number;
};