# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.4.0
  description: Oogly Chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Messages
...
  /bans:
    get:
      operationId: listBans
      tags: [bans]
      summary: List all bans (admin)
      parameters:
        - in: query
          name: active
          schema:
            type: boolean
      responses:
        "200":
          description: Bans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ban"

    post:
      operationId: banUser
      tags: [bans]
      summary: Ban a user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanInput"
      responses:
        "201":
          description: Banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ban"
  /bans/{id}:
    delete:
      operationId: unbanUser
      tags: [bans]
      summary: Remove a ban (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unbanned
  /users/{id}/ban-status:
    get:
      operationId: getUserBanStatus
      tags: [bans]
      summary: Get ban status for a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ban status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BanStatus"

  # Mutes
...
  /appeals:
    get:
      operationId: listAppeals
      tags: [appeals]
      summary: List all appeals (admin)
      parameters:
        - in: query
          name: unread
          schema:
            type: boolean
      responses:
        "200":
          description: Appeals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appeal"

    post:
      operationId: submitAppeal
      tags: [appeals]
      summary: Submit an appeal (banned/muted user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppealInput"
      responses:
        "201":
          description: Appeal submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appeal"
  /appeals/{id}/read:
    patch:
      operationId: markAppealRead
      tags: [appeals]
      summary: Mark appeal as read (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appeal"
  /appeals/{id}/dismiss:
    patch:
      operationId: dismissAppeal
      tags: [appeals]
      summary: Dismiss an appeal (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appeal"

  # Announcements
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"

    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Moderation Logs
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # User
    User:
      type: object
      required: [id, username, isAdmin, isOwner, isMuted, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        isAdmin:
          type: boolean
        isOwner:
          type: boolean
        isMuted:
          type: boolean
        theme:
          type: ["string", "null"]
        notificationsEnabled:
          type: ["boolean", "null"]
        soundsEnabled:
          type: ["boolean", "null"]
        compactMode:
          type: ["boolean", "null"]
        lastSeen:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Ban:
      type: object
      required: [id, userId, reason, isPermanent, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: ["string", "null"]
        bannedById:
          type: ["integer", "null"]
        bannedByUsername:
          type: ["string", "null"]
        reason:
          type: string
        isPermanent:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    BanInput:
      type: object
      required: [userId, reason]
      properties:
        userId:
          type: integer
        reason:
          type: string
          minLength: 1
        isPermanent:
          type: boolean
        expiresAt:
          type: ["string", "null"]
    BanStatus:
      type: object
      required: [isBanned]
      properties:
        isBanned:
          type: boolean
        ban:
          $ref: "#/components/schemas/Ban"

    # Mutes
...
    Appeal:
      type: object
      required: [id, userId, message, isRead, isDismissed, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: ["string", "null"]
        message:
          type: string
        isRead:
          type: boolean
        isDismissed:
          type: boolean
        createdAt:
          type: string
    AppealInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 10
          maxLength: 2000

    # Announcements
    Announcement:
      type: object
      required: [id, content, authorId, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: ["string", "null"]
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000

    # Moderation Logs
...