# lib/api-zod/src/generated/types/appealInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oogly Chat API specification
 * OpenAPI spec version: 1.4.0
 */

export interface AppealInput {
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
}