# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oogly Chat API specification
 * OpenAPI spec version: 1.4.0
 */

export interface User {
  id: number;
  username: string;
  isAdmin: boolean;
  isOwner: boolean;
  isMuted: boolean;
  /** @nullable */
  theme?: string | null;
  /** @nullable */
  notificationsEnabled?: boolean | null;
  /** @nullable */
  soundsEnabled?: boolean | null;
  /** @nullable */
  compactMode?: boolean | null;
  /** @nullable */
  lastSeen?: string | null;
  createdAt: string;
}